requirements.txt
fastapi
uvicorn
streamlit
google-api-python-client
google-auth
google-auth-oauthlib
langchain
langgraph
chromadb
numpy
